+++ Films.Infrastructure.Web/Comments/ViewModels/CommentsViewModel.cs
@@ -3,5 +3,5 @@ namespace Films.Infrastructure.Web.Comments.ViewModels;
 public class CommentsViewModel
 {
     public required IEnumerable<CommentViewModel> Comments { get; init; }
-    public required long CountPages { get; init; }
+    public required int CountPages { get; init; }
 }
